modularized the controller files and updated routes
